refactor(iroh-dns): extract system DNS reader into a private module
Move the body of HickoryResolver::system_config into a new
crate::system_config module with a structured error type. No behavior
change. Sets up follow-on commits that need a single platform-cfg
choke point for Android-specific defenses.
