+++ iroh-dns/src/dns.rs
@@ -600,7 +600,7 @@ impl HickoryResolver {
 
     fn build_resolver(builder: &Builder) -> TokioResolver {
         let (mut config, mut options) = if builder.use_system_defaults {
-            match Self::system_config() {
+            match crate::system_config::read_system_conf() {
                 Ok((config, options)) => (config, options),
                 Err(error) => {
                     debug!(%error, "Failed to read the system's DNS config, using fallback DNS servers.");
@@ -637,26 +637,6 @@ impl HickoryResolver {
         }
 
         hickory_builder.build().expect("config works")
-    }
-
-    fn system_config() -> Result<(ResolverConfig, ResolverOpts), hickory_resolver::net::NetError> {
-        let (system_config, options) = hickory_resolver::system_conf::read_system_conf()?;
-
-        // Copy all of the system config, but strip the bad windows nameservers.  Unfortunately
-        // there is no easy way to do this.
-        let mut config = hickory_resolver::config::ResolverConfig::default();
-        if let Some(name) = system_config.domain() {
-            config.set_domain(name.clone());
-        }
-        for name in system_config.search() {
-            config.add_search(name.clone());
-        }
-        for nameserver_cfg in system_config.name_servers() {
-            if !WINDOWS_BAD_SITE_LOCAL_DNS_SERVERS.contains(&nameserver_cfg.ip) {
-                config.add_name_server(nameserver_cfg.clone());
-            }
-        }
-        Ok((config, options))
     }
 }
 
@@ -765,20 +745,6 @@ impl From<Vec<Box<[u8]>>> for TxtRecordData {
         Self(value.into_boxed_slice())
     }
 }
-
-/// Deprecated IPv6 site-local anycast addresses still configured by windows.
-///
-/// Windows still configures these site-local addresses as soon even as an IPv6 loopback
-/// interface is configured.  We do not want to use these DNS servers, the chances of them
-/// being usable are almost always close to zero, while the chance of DNS configuration
-/// **only** relying on these servers and not also being configured normally are also almost
-/// zero.  The chance of the DNS resolver accidentally trying one of these and taking a
-/// bunch of timeouts to figure out they're no good are on the other hand very high.
-const WINDOWS_BAD_SITE_LOCAL_DNS_SERVERS: [IpAddr; 3] = [
-    IpAddr::V6(Ipv6Addr::new(0xfec0, 0, 0, 0xffff, 0, 0, 0, 1)),
-    IpAddr::V6(Ipv6Addr::new(0xfec0, 0, 0, 0xffff, 0, 0, 0, 2)),
-    IpAddr::V6(Ipv6Addr::new(0xfec0, 0, 0, 0xffff, 0, 0, 0, 3)),
-];
 
 /// Helper enum to give a unified type to the iterators of [`DnsResolver::lookup_ipv4_ipv6`].
 enum LookupIter<A, B> {
